refactor utility functions to new class
