+++ plugins/yospos.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 
 from rtmbot.core import Plugin
+from .util import Util as util
 import requests
 import json
 import configparser
@@ -15,7 +16,7 @@ class WeatherPlugin(Plugin):
     """
     config = configparser.ConfigParser()
     config.read('weather.conf')
-    dark_sky_key = config['default']['DarkSkyKey']
+    dark_sky_key = config['default']['DarkSkyKey'] if config['default']['DarkSkyKey'] else "none"
 
     def process_message(self, msg):
         if check_message(msg):
@@ -30,7 +31,7 @@ class WeatherPlugin(Plugin):
             self.dark_sky_key, coords['lat'], coords['lon']),
             params={'units': 'si'})
         response = json.loads(r.text)
-        message = format_weather_message(coords['display_name'], str(response['currently']['temperature']), response['currently']['summary'])
+        message = util.format_weather_message(coords['display_name'], str(response['currently']['temperature']), response['currently']['summary'])
         return json.dumps(message)
 
 def get_coordinates(loc):
@@ -41,36 +42,3 @@ def get_coordinates(loc):
 
 def check_message(msg):
     return msg['text'].startswith('.wea')
-
-def c_to_f(temp):
-    return format((float(temp) * 1.8) + 32, '.1f')
-
-def format_weather_message(loc_string, temp_string, summary_string):
-    response = {
-        'attachments': [
-            {
-                'fallback': "Weather in {0} - {1}°C / {2}°F".format(loc_string, temp_string, c_to_f(temp_string)),
-                'pretext': "I found some weather, y'all...",
-                'color': '#57FF57',
-                'text': "Current weather for {0}".format(loc_string),
-                'fields': [
-                    {
-                        'title': "Celcius",
-                        'value': "{0}".format(temp_string),
-                        'short': True
-                    },
-                    {
-                        'title': "Farenheit",
-                        'value': "{0}".format(c_to_f(temp_string)),
-                        'short': True
-                    },
-                    {
-                        'title': "Summary",
-                        'value': "{0}".format(summary_string),
-                        'short': False
-                    }
-                ]
-            }
-        ]
-    }
-    return response
